Add: terrain diverse spawn Fix: terrain resource positioning

+++ src/main/java/ap/project/civilization/view/ui/Camera.java
@@ -31,10 +31,10 @@ public class Camera {
         centerOn((int)k.getX(), (int)k.getY());
     }
 
-    public int worldToScreenX(int x) {
+    public double worldToScreenX(double x) {
         return x - screenX;
     }
-    public int worldToScreenY(int y) {
+    public double worldToScreenY(double y) {
         return y - screenY;
     }
